Split do_import into its own file
