+++ corehq/apps/case_importer/management/commands/import_cases.py
@@ -5,7 +5,7 @@ import json
 from datetime import datetime
 from django.core.management import BaseCommand, CommandError
 from dimagi.utils.web import json_handler
-from corehq.apps.case_importer.tasks import do_import
+from corehq.apps.case_importer.do_import import do_import
 from corehq.apps.case_importer.util import ImporterConfig, get_spreadsheet
 from corehq.apps.users.models import WebUser
 from io import open
